backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
psycopg-binary==3.2.3
python-decouple==3.8
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
Pillow==11.0.0
django-filter==23.5
